requirements.txt
numpy
matplotlib
pillow
ipython
